Move the refusal log formatter into src/safe-log.js
external-url.js and site-registry.js each carried their own copy of the
escape-and-truncate step that renders a refused value into the log file,
and site-registry.js's copy said in a comment that the third caller
should be the one to move it somewhere shared rather than add a third.

The editor-launch module for #150 is that third caller, so this moves it
first: safe-log.js holds `describeRefused` and the reasoning, and the two
existing formatters become one-line wrappers over it so their names — and
their tests, unchanged — still say which guard is refusing.

Behaviour is identical; test/external-url.test.cjs and
test/site-registry.test.cjs pass untouched, which is the point.
